more emacs 31

+++ linkapps.nix
@@ -16,7 +16,11 @@
       shopt -s nullglob
       for app in "$newGenPath"/home-path{,/Applications}/*.app; do
         app="$(realpath "$app")"
-        targetApp="$targetDir/$(basename "$app")"
+        name="$(basename "$app")"
+        # Emacs.app is installed as a real, code-signed copy by
+        # home.activation.signEmacsApp; an alias here would clash with it.
+        if [[ $name == Emacs.app ]]; then continue; fi
+        targetApp="$targetDir/$name"
         if [[ -f $targetApp ]]; then run rm "$targetApp"; fi
         run ${pkgs.mkalias}/bin/mkalias -v "$app" "$targetApp"
       done
